requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
pydantic==2.7.3
opencv-python
